some late touchups for sparse
Add an example showing how to interop with the 'sprs' crate.
(which has like 50 deps including a public dependency on ndarray and
 we would never ever provide a direct conversion function)

Allow the sparse types to take borrowed data.  This makes the entire
module look obscenely more complicated.  But it's also the kind of
generic trash many of us Rustaceans are used to seeing by this point.

Fix a bug accidentally using File::open instead of File::create >_>