add zip to dev-dependencies
There was some funniness in Cargo.toml to allow one of the tests
to enable one of the optional dependencies  (for use inside the
test, rather than the crate), with the provided justification:

    only required because an optional dependency can't also be a dev dependency

But, uh..... I can't find anything to substantiate this. Adding zip to
dev-dependencies seems to do exactly what's expected ('cargo build'
does not depend on zip, while 'cargo test' and 'cargo run --example EXAMPLE'
does).  Even back in, like, cargo 1.40.0.  So I'm not sure why I wrote that.